Make the behavior consistent when setting image pull secrets for both the Deployment and StatefulSet
For example, when setting 'spec.application.imagePullSecrets' to an empty list,
the behavior was not consistent: for Backstage, the resulting pull secrets was empty,
and the DB stateful would inherit the pull secrets coming from the default config.
This makes the behavior consistent by overriding the pull secrets in the pod spec
as long as 'spec.application.imagePullSecrets' is set.
If not set, we inherit the pull secrets from the default config.