[1.1.x] share configured Image Pull Secret to the Database pod to make it possible to pull non-public PSQL image (#290)
* share image-pull secret for database

* fix Postgre description

* Revert change to bundle/manifests/backstage-operator.clusterserviceversion.yaml

* Update controllers/backstage_controller_test.go

* Make the behavior consistent when setting image pull secrets for both the Deployment and StatefulSet

For example, when setting 'spec.application.imagePullSecrets' to an empty list,
the behavior was not consistent: for Backstage, the resulting pull secrets was empty,
and the DB stateful would inherit the pull secrets coming from the default config.
This makes the behavior consistent by overriding the pull secrets in the pod spec
as long as 'spec.application.imagePullSecrets' is set.
If not set, we inherit the pull secrets from the default config.

---------

Co-authored-by: Armel Soro <armel@rm3l.org>
Co-authored-by: Armel Soro <asoro@redhat.com>