Handle patch error and delete StatefulSet so it can be recreated
Some resources like StatefulSets allow patching a limited set of fields.
For cases where we are upgrading existing instances, we might need to
change the StatefulSet spec.
The recommended approach to doing this is to delete the existing
StatefulSet, but keep its dependents (like PVCs and Pods) orphan, then
recreate the StatefulSet.
This way, it will be upgraded (with some downtime) with no data loss.