Generalize the logic of deleting the resource if the patching operation fails
This way, it will be recreated at the next reconciliation.
This should prevent similar issues from happening again in the future.

The DB Service could not be patched to set its clusterIP field to None (to make it headless),
when we were upgrading from 1.1 (where it was not headless) to 1.2.

In fact, regarding K8s Services, users can upgrade (add another clusterIP or ipFamily),
downgrade (remove secondary clusterIP or ipFamily),
but *CAN NOT* change primary/secondary clusterIP || ipFamily
*UNLESS* they are changing from/to/ON ExternalName