[1.2.x] fix: Fix reconciliation error preventing existing Backstage operands from being upgraded when the operator is upgraded from `1.1.x` [RHIDP-2597] (#385)
* Add E2E test highlighting the issue

* Run upgrade path E2E test with dedicated filter

This is because this test requires uninstalling the installed operator, in order to install a previous version, then try out the upgrade

* Run the upgrade path E2E tests in the Nightly job

* Generalize the logic of deleting the resource if the patching operation fails

This way, it will be recreated at the next reconciliation.
This should prevent similar issues from happening again in the future.

The DB Service could not be patched to set its clusterIP field to None (to make it headless),
when we were upgrading from 1.1 (where it was not headless) to 1.2.

In fact, regarding K8s Services, users can upgrade (add another clusterIP or ipFamily),
downgrade (remove secondary clusterIP or ipFamily),
but *CAN NOT* change primary/secondary clusterIP || ipFamily
*UNLESS* they are changing from/to/ON ExternalName

* fixup! Add E2E test highlighting the issue

* Unconditionally try to delete the object (so it can be recreated) if patching did not succeed

Co-authored-by: Gennady Azarenkov <gazarenkov@gmail.com>

* fixup! Unconditionally try to delete the object (so it can be recreated) if patching did not succeed

* Move upgrade-related tests to a specific spec file

Co-authored-by: Gennady Azarenkov <gazarenkov@gmail.com>

* Call out in the logs that manually-added labels/annotations will be lost when the object is deleted

Co-authored-by: Gennady Azarenkov <gazarenkov@gmail.com>

---------

Co-authored-by: Armel Soro <asoro@redhat.com>
Co-authored-by: Gennady Azarenkov <gazarenkov@gmail.com>