Switch the default operand image tag from 'latest' to 'next'
'next' is the development tag of the upstream Showcase image,
and should hopefully be stable.
The goal with this is to monitor how our tests behave against this 'next' tag. If it proves too unstable,
we might want to switch back to using the downstream RHDH image
(`quay.io/rhdh/rhdh-hub-rhel9:latest`).